+++ R/1.3-exprso.R
@@ -29,7 +29,11 @@
 #'  \code{class(y) == "character"} or \code{class(y) == "factor"},
 #'  \code{exprso} prepares data for binary or multi-class classification.
 #'  Else, \code{exprso} prepares data for regression. If \code{y} is a
-#'  matrix, the program assumes the first column is the outcome.
+#'  matrix, the program uses the column in \code{label}.
+#' @param label A numeric scalar or character string. The column to
+#'  use as the label if \code{y} is a matrix.
+#' @param switch A logical scalar. Toggles which class label is
+#'  called Control in binary classification.
 #' @return An \code{ExprsArray} object.
 #'
 #' @examples
@@ -44,9 +48,9 @@
 #' predict(mach, arrays[[2]])
 #' }
 #' @export
-exprso <- function(x, y){
+exprso <- function(x, y, label = 1, switch = FALSE){
 
-  if(length(y) != nrow(x)) stop("Incorrect number of outcomes.")
+  if(length(label) > 1) stop("More than one label specified.")
   array <-
     new("ExprsArray",
         exprs = t(as.data.frame(x)), annot = as.data.frame(y),
@@ -57,21 +61,19 @@ exprso <- function(x, y){
   colnames(array@exprs) <- paste0("x", 1:ncol(array@exprs))
   colnames(array@exprs) <- make.names(colnames(array@exprs), unique = TRUE)
   rownames(array@annot) <- colnames(array@exprs)
-  labels <- array@annot[,1]
+  labels <- array@annot[,label]
 
   # Set sub-class to guide fs and build modules
+  if(length(labels) != nrow(x)) stop("Incorrect number of outcomes.")
   if(class(labels) == "logical") stop("Boolean outcomes not supported.")
   if(class(labels) == "character" | class(labels) == "factor"){
     if(length(unique(y)) == 2){
       print("Preparing data for binary classification.")
       class(array) <- "ExprsBinary"
-      if(all(unique(labels) %in% c("Control", "Case"))){
-        print("Using labels provided as CONTROL / CASE.")
-        array@annot$defineCase <- as.character(labels)
-      }else{
-        print("Converting labels to CONTROL / CASE.")
-        array@annot$defineCase <- ifelse(labels == unique(labels)[1], "Control", "Case")
-      }
+      print("Converting binary labels to CONTROL / CASE.")
+      control <- unique(labels)[as.numeric(switch) + 1]
+      print(paste("CONTROL:", control, "(override with 'switch')"))
+      array@annot$defineCase <- ifelse(labels == control, "Control", "Case")
     }else{
       print("Preparing data for multi-class classification.")
       class(array) <- "ExprsMulti"